fix bugs with sort by created time, and bug with check course in search page

+++ lib/models/video_model/lesson_model.dart
@@ -35,7 +35,7 @@ class LessonModel {
   LessonModel.fromJson(Map<String, dynamic> json) {
     link = json['link'];
     name = json['name'];
-    duration = json['duration'];
+    duration = json['duration'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
